backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pdfplumber>=0.11.0
httpx>=0.27.0
redis>=5.0.0
pydantic>=2.8.0
python-dotenv>=1.0.1
pytest>=8.0.0
